Refactor to seperate the 'table' view of addresses for the radius search, and the results generated for the rest of the pages
