+++ src/aims_ui/static/js/a_single_matches/radiussearch/interactive_map/map_setup.mjs
@@ -1,4 +1,4 @@
-import { replaceDistancePlaceholderWithSearchValues } from '/static/js/macros/custom_address_info/distance_calculator.mjs';
+import { replaceDistancePlaceholderWithSearchValues } from '/static/js/macros/custom_address_results_table/distance_calculator.mjs';
 import { makePinIcon } from '/static/js/a_single_matches/radiussearch/interactive_map/map_pin_icons.mjs';
 import { setMapZoomInPageStorage,  getCurrentSearchLatValue, getCurrentSearchLonValue, getStartLatValue, getStartLonValue, getStartZoomValue } from './input_and_stored_values.mjs';
 import { getDefaultValuesForPage } from '/static/js/e_all_pages/setup_defaults.mjs';
