+++ src/aims_ui/templates/aims_ui_pages/a_single_matches/radiussearch.html
@@ -6,7 +6,7 @@
 {% from "components/section-navigation/_macro.njk" import onsSectionNavigation %}
 
 {% from 'custom_search/_macro.njk' import customSearch %}
-{% from 'custom_address_info/_macro.njk' import customAddressInfo %}
+{% from 'custom_address_results_table/_macro.njk' import customAddressResultsTable %}
 {% from 'custom_save_addresses_to_client/_macro.njk' import customSaveAddressesToClient %}
 {% from 'custom_we_matched_message/_macro.njk' import weMatchedAddressPreamble %}
 {% from 'custom_title_and_description/_macro.njk' import customTitleAndDescription %}
@@ -73,10 +73,9 @@
 					"page_name": page_name,
 				}) }}
 
-				{{ customAddressInfo ({
+				{{ customAddressResultsTable({
 					'matched_addresses': matched_addresses,
 					'page_name': page_name,
-					'compressed_view': True,
 					'results_page': results_page,
 					}) }}
 			</div>
